Recorded response time for answers
In the room object I added an extra dict attribute called responseTimes which measures the time remaining of the question before it was answered. It is indexed by socked id and question Index similar to the answers attribute dict. Time taken to answer is calculated in the handle answer socket.io listener function and is passed onto the handleAnswer function to calculate the time remaining which is stored in responseTimes.