QR Code on WaitLobby and Podium on gameEnd_screen (in progress)
A QR code is visible on the waitLobby of the screen that navigates to docs/[slug] page of our deployed JAM website and saves the roomid in the $user.gameid so that a player does not have to manually enter the room number. Then it navigates to the landing page of the JAM website.

Podium on the gameEnd_ screen is a work in progress